Refactor Dataset to reduce file count
