+++ src/dataset/index.js
@@ -1,8 +1,17 @@
 var extend = require("../core/utils/extend"),
     Dataset = require("./dataset");
 
-extend(Dataset.prototype, {
+extend(Dataset.prototype, require("./lib/append"));
+extend(Dataset.prototype, require("./lib/filter"));
+extend(Dataset.prototype, require("./lib/insert"));
+extend(Dataset.prototype, require("./lib/remove"));
+extend(Dataset.prototype, require("./lib/select"));
+extend(Dataset.prototype, require("./lib/sort"));
+extend(Dataset.prototype, require("./lib/update"));
 
+extend(Dataset.prototype, require("./lib/analyses"));
+extend(Dataset.prototype, {
+  "format": require("./lib/format"),
 });
 
 module.exports = Dataset;
