+++ src/dataset/lib/append.js
@@ -1,0 +1,77 @@
+var each = require("../../core/utils/each");
+
+module.exports = {
+  "appendColumn": appendColumn,
+  "appendRow": appendRow
+};
+
+function appendColumn(str, input){
+  var self = this,
+      args = Array.prototype.slice.call(arguments, 2),
+      label = (str !== undefined) ? str : null;
+
+  if (typeof input === "function") {
+    self.data.output[0].push(label);
+    each(self.output(), function(row, i){
+      var cell;
+      if (i > 0) {
+        cell = input.call(self, row, i);
+        if (typeof cell === "undefined") {
+          cell = null;
+        }
+        self.data.output[i].push(cell);
+      }
+    });
+  }
+
+  else if (!input || input instanceof Array) {
+    self.data.output[0].push(label);
+    each(self.output(), function(row, i){
+      var cell;
+      if (i > 0) {
+        cell = (input && input[i-1] !== undefined) ? input[i-1] : null;
+        self.data.output[i].push(cell);
+      }
+    });
+
+  }
+
+  return self;
+}
+
+function appendRow(str, input){
+  var self = this,
+      args = Array.prototype.slice.call(arguments, 2),
+      label = (str !== undefined) ? str : null,
+      newRow = [];
+
+  newRow.push(label);
+
+  if (typeof input === "function") {
+    each(self.output()[0], function(label, i){
+      var col, cell;
+      if (i > 0) {
+        col = self.selectColumn(i);
+        cell = input.call(self, col, i);
+        if (typeof cell === "undefined") {
+          cell = null;
+        }
+        newRow.push(cell);
+      }
+    });
+    self.data.output.push(newRow);
+  }
+
+  else if (!input || input instanceof Array) {
+    each(self.output()[0], function(label, i){
+      var cell;
+      if (i > 0) {
+        cell = (input && input[i-1] !== undefined) ? input[i-1] : null;
+        newRow.push(cell);
+      }
+    });
+    this.data.output.push(newRow);
+  }
+
+  return this;
+}
