--- src/dataset/lib/removeColumn.js
@@ -1,11 +1,0 @@
-Keen.Dataset.prototype.deleteColumn = function(q){
-  var self = this,
-      index = (!isNaN(parseInt(q))) ? q : this.output()[0].indexOf(q);
-
-  if (index > -1) {
-    each(self.data.output, function(row, i){
-      self.data.output[i].splice(index, 1);
-    });
-  }
-  return self;
-};
