+++ src/dataset/lib/sort.js
@@ -1,0 +1,57 @@
+var each = require("../../core/utils/each");
+
+module.exports = {
+  "sortColumns": sortColumns,
+  "sortRows": sortRows
+};
+
+function sortColumns(str, comp){
+  var self = this,
+      head = this.output()[0].slice(1), // minus index
+      cols = [],
+      clone = [],
+      fn = comp || this.getColumnLabel;
+
+  // Isolate each column (except the index)
+  each(head, function(cell, i){
+    cols.push(self.selectColumn(i+1).slice(0));
+  });
+  cols.sort(function(a,b){
+    // If fn(a) > fn(b)
+    var op = fn.call(self, a) > fn.call(self, b);
+    if (op) {
+      return (str === "asc" ? 1 : -1);
+    } else if (!op) {
+      return (str === "asc" ? -1 : 1);
+    } else {
+      return 0;
+    }
+  });
+  each(cols, function(col, i){
+    self
+    .deleteColumn(i+1)
+    .insertColumn(i+1, col[0], col.slice(1));
+  });
+  return self;
+}
+
+function sortRows(str, comp){
+  var self = this,
+      head = this.output().slice(0,1),
+      body = this.output().slice(1),
+      fn = comp || this.getRowIndex;
+
+  body.sort(function(a, b){
+    // If fn(a) > fn(b)
+    var op = fn.call(self, a) > fn.call(self, b);
+    if (op) {
+      return (str === "asc" ? 1 : -1);
+    } else if (!op) {
+      return (str === "asc" ? -1 : 1);
+    } else {
+      return 0;
+    }
+  });
+  self.output(head.concat(body));
+  return self;
+}
